refactor : product model-domain 분리
