+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/product/dto/ProductResponseDto.java
@@ -1,6 +1,6 @@
 package com.sparta.blackwhitedeliverydriver.product.dto;
 
-import com.sparta.blackwhitedeliverydriver.product.entity.Product;
+import com.sparta.blackwhitedeliverydriver.model.product.entity.Product;
 import java.util.UUID;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
