+++ blackWhiteDeliveryDriver/src/main/java/com/sparta/blackwhitedeliverydriver/product/service/ProductService.java
@@ -4,16 +4,16 @@ import com.sparta.blackwhitedeliverydriver.category.dto.CreateProductRequestDto;
 import com.sparta.blackwhitedeliverydriver.product.dto.ProductIdResponseDto;
 import com.sparta.blackwhitedeliverydriver.product.dto.ProductRequestDto;
 import com.sparta.blackwhitedeliverydriver.product.dto.ProductResponseDto;
-import com.sparta.blackwhitedeliverydriver.product.entity.Product;
-import com.sparta.blackwhitedeliverydriver.store.entity.Store;
-import com.sparta.blackwhitedeliverydriver.user.entity.User;
-import com.sparta.blackwhitedeliverydriver.user.entity.UserRoleEnum;
+import com.sparta.blackwhitedeliverydriver.model.product.entity.Product;
+import com.sparta.blackwhitedeliverydriver.model.store.entity.Store;
+import com.sparta.blackwhitedeliverydriver.model.user.entity.User;
+import com.sparta.blackwhitedeliverydriver.model.user.entity.UserRoleEnum;
 import com.sparta.blackwhitedeliverydriver.user.exception.ExceptionMessage;
 import com.sparta.blackwhitedeliverydriver.product.exception.ProductExceptionMessage;
 import com.sparta.blackwhitedeliverydriver.store.exception.StoreExceptionMessage;
-import com.sparta.blackwhitedeliverydriver.product.repository.ProductRepository;
-import com.sparta.blackwhitedeliverydriver.store.repository.StoreRepository;
-import com.sparta.blackwhitedeliverydriver.user.repository.UserRepository;
+import com.sparta.blackwhitedeliverydriver.model.product.repository.ProductRepository;
+import com.sparta.blackwhitedeliverydriver.model.store.repository.StoreRepository;
+import com.sparta.blackwhitedeliverydriver.model.user.repository.UserRepository;
 import com.sparta.blackwhitedeliverydriver.common.security.UserDetailsImpl;
 import java.time.LocalDateTime;
 import java.util.Optional;
